move UOp add_chain, mod_folding, div_folding into cached methods
is this faster?
[run_process_replay]
